massive work, doubting choices already

+++ generic/print.h
@@ -12,9 +12,9 @@
 struct print_table
 {
     vector<uint64> column_widths_ = {};
-    uint16 table_width_ = 0;
+    uint32 table_width_ = 0;
     const string field_name_header_ = "Field name";
-    uint16 field_name_width_ = field_name_header_.length();
+    uint32 field_name_width_ = field_name_header_.length();
     vector<string> field_names_ = {};
     int64 current_row_ = -1;
 
@@ -123,7 +123,7 @@ struct print_table
         return std::to_string(input);
     }
 
-    uint16 calculate_table_width()
+    uint32 calculate_table_width()
     {
         return 1 +
                 field_name_width_ + 1 +
@@ -134,7 +134,7 @@ struct print_table
 
     void print_hor_line()
     {
-        for(uint16 i = 1; i <= table_width_; i++)
+        for(uint32 i = 1; i <= table_width_; i++)
         {
             std::cout << "-";
         }
@@ -151,7 +151,7 @@ struct print_table
     {
         std::cout << "|" << str;
 
-        for(uint16 i = 0; i < table_width_ - str.length() - 2; i++)
+        for(uint32 i = 0; i < table_width_ - str.length() - 2; i++)
         {
             std::cout << " ";
         }
